pkg/processor/runtime/python/py/requirements/common.txt
nuclio-sdk
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-prometheus